+++ src/main/java/frc/robot/RobotContainer.java
@@ -3,14 +3,19 @@
 // the WPILib BSD license file in the root directory of this project.
 
 package frc.robot;
-
-
-import java.util.ResourceBundle.Control;
-
 import edu.wpi.first.wpilibj.smartdashboard.SendableChooser;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
-import frc.robot.Trajectories.TrajectoryCreation;
-import frc.robot.commands.*;
+import frc.robot.commands.Climb.ExtendClimb;
+import frc.robot.commands.Climb.Pivot;
+import frc.robot.commands.Climb.RetractClimb;
+import frc.robot.commands.Climb.ToggleHooks;
+import frc.robot.commands.Drivetrain.DefaultDrive;
+import frc.robot.commands.Drivetrain.FollowTrajectory;
+import frc.robot.commands.Drivetrain.TrajectoryCreation;
+import frc.robot.commands.Intake.ArmForward;
+import frc.robot.commands.Intake.ArmReverse;
+import frc.robot.commands.Intake.BottomClose;
+import frc.robot.commands.Intake.BottomOpen;
 import frc.robot.controls.ControlMap;
 import frc.robot.subsystems.Climb;
 import frc.robot.subsystems.Drivetrain;
@@ -25,43 +30,20 @@ import edu.wpi.first.wpilibj2.command.Command;
  * subsystems, commands, and button mappings) should be declared here.
  */
 public class RobotContainer {
-  // The robot's subsystems and commands are defined here...
+
+  //Subsystem Declarations
   private final Climb m_climb = new Climb();
-  
-  private final Pivot m_autoCommand = new Pivot(m_climb);
-
-  private final TrajectoryCreation m_trajectory = new TrajectoryCreation();
-
   private final Intake m_intake = new Intake();
-  private final ExtendClimb m_extend = new ExtendClimb(m_climb);
-  private final RetractClimb m_retract = new RetractClimb(m_climb);
-  private final HookOff m_hookOff = new HookOff(m_climb);
-  private final HookOn m_hookOn = new HookOn(m_climb);
-  private final Pivot m_pivot = new Pivot(m_climb);
-  private final ToggleClimb m_toggleClimb = new ToggleClimb(m_climb);
-  private final ToggleHooks m_toggleHooks = new ToggleHooks(m_climb);
-  private final TopClose m_topclose = new TopClose(m_intake);
-  private final TopOpen m_topopen = new TopOpen(m_intake);
-  private final BottomClose m_bottomclose = new BottomClose(m_intake);
-  private final BottomOpen m_bottomopen = new BottomOpen(m_intake);
-  private final SendableChooser<Command> m_chooser = new SendableChooser<>();
   private final Drivetrain m_drivetrain = new Drivetrain();
-  private final DefaultDrive m_default = new DefaultDrive(m_drivetrain);
 
+  //configure default commands
+  private final DefaultDrive m_default = new DefaultDrive(m_drivetrain);
+  private final Pivot m_pivot = new Pivot(m_climb);
+
+  //Trajectories
   private final FollowTrajectory m_follower = new FollowTrajectory();
   private final TrajectoryCreation m_traj = new TrajectoryCreation();
-
-  //private final RotateWristDown m_rotatewristdown = new RotateWristDown(m_intake);
-  //private final RotateWristUp m_rotatewristup = new RotateWristUp(m_intake);
-  private final ArmForward m_arm = new ArmForward(m_intake);
-  //private final StartEndCommand m_start_end_top_servo = new StartEndCommand(m_intake::TopServoClose, m_intake::TopServoOpen, m_intake);
-  //private final StartEndCommand m_start_end_bottom_servo = new StartEndCommand(m_intake::BottomServoClose, m_intake:: BottomServoOpen, m_intake);
-  //private final StartEndCommand m_start_end_wrist_servo = new StartEndCommand(m_intake::WristClose, m_intake::WristOpen, m_intake);
-
-
-  /**
-   * The container for the robot. Contains subsystems, OI devices, and commands.
-   */
+  private final SendableChooser<Command> m_chooser = new SendableChooser<>();
 
   public RobotContainer() {
     // Configure the button bindings
@@ -76,8 +58,8 @@ public class RobotContainer {
    * edu.wpi.first.wpilibj2.command.button.JoystickButton}.
    */
   private void configureButtonBindings() {
-    m_chooser.addOption("Mecanum Trajectory", m_follower.generateTrajectory(m_drivetrain, m_trajectory.testTrajectory));
-    m_chooser.addOption("Bill", m_follower.generateTrajectory(m_drivetrain, m_trajectory.testTrajectory2));
+    m_chooser.addOption("Mecanum Trajectory", m_follower.generateTrajectory(m_drivetrain, m_traj.testTrajectory));
+    m_chooser.addOption("Bill", m_follower.generateTrajectory(m_drivetrain, m_traj.testTrajectory2));
     //m_chooser.addOption("Manual Code", object);
     m_chooser.addOption("Line up 1v1", m_follower.generateTrajectory(m_drivetrain, m_traj.path1v1));
     m_chooser.addOption("Line up 1v2", m_follower.generateTrajectory(m_drivetrain, m_traj.path1v2));
@@ -98,20 +80,9 @@ public class RobotContainer {
 
     ControlMap.climbExtend.whenPressed(new ExtendClimb(m_climb));
     ControlMap.climbRetract.whenPressed(new RetractClimb(m_climb));
-    ControlMap.staticHookOn.toggleWhenPressed(m_toggleHooks);
+    ControlMap.staticHookOn.toggleWhenPressed(new ToggleHooks(m_climb));
     ControlMap.arm_down.whenPressed(new ArmForward(m_intake));
     ControlMap.arm_down.whenPressed(new ArmReverse(m_intake));
-    
-
-    //ControlMap.climbExtend.toggleWhenPressed(m_toggleClimb);
-
-    // ControlMap.staticHookOn.whenPressed(new HookOn(m_climb));
-    // ControlMap.staticHookOff.whenPressed(new HookOff(m_climb));
-    
-    
-    /*ControlMap.topToggle.toggleWhenPressed(m_start_end_top_servo);
-    ControlMap.bottomToggle.toggleWhenPressed(m_start_end_bottom_servo);
-    ControlMap.wristToggle.toggleWhenPressed(m_start_end_wrist_servo);*/
     
   }
