+++ src/main/java/frc/robot/commands/Climb/ExtendClimb.java
@@ -2,35 +2,30 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
+package frc.robot.commands.Climb;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
 import frc.robot.subsystems.Climb;
 
-public class ToggleHooks extends CommandBase {
-  /** Creates a new ToggleHooks. */
+public class ExtendClimb extends CommandBase {
   private final Climb m_climb;
-  public ToggleHooks(Climb climb) {
+
+  //Constructor
+  public ExtendClimb(Climb climb){
     m_climb = climb;
-    // Use addRequirements() here to declare subsystem dependencies.
     addRequirements(climb);
   }
-
-  // Called when the command is initially scheduled.
+  //Extend the pivot arm
   @Override
-  public void initialize() {
-    m_climb.ServoOpen();
+  public void initialize(){
+    m_climb.extendArm();
   }
-
-  // Called every time the scheduler runs while the command is scheduled.
   @Override
   public void execute() {}
 
   // Called once the command ends or is interrupted.
   @Override
-  public void end(boolean interrupted) {
-    m_climb.ServoClose();
-  }
+  public void end(boolean interrupted) {}
 
   // Returns true when the command should end.
   @Override
