+++ src/main/java/frc/robot/commands/Climb/Pivot.java
@@ -2,45 +2,46 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
+package frc.robot.commands.Climb;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
-import frc.robot.subsystems.Intake;
+import frc.robot.controls.ControlMap;
+import frc.robot.subsystems.Climb;
 
-public class ArmReverse extends CommandBase {
-  /** Creates a new TopClose. */
-  
-  private final Intake m_intake;
-  public ArmReverse(Intake intake) {
+
+public class Pivot extends CommandBase {
+  /** Creates a new Pivot. */
+  private final Climb m_pivot;
+  public Pivot(Climb climb) {
     // Use addRequirements() here to declare subsystem dependencies.
-    m_intake = intake;
-    addRequirements(intake);
+    m_pivot = climb;
+    addRequirements(climb);
+
   }
 
   // Called when the command is initially scheduled.
   @Override
   public void initialize() {
-    
+    m_pivot.ServoClose();
+
   }
 
   // Called every time the scheduler runs while the command is scheduled.
   @Override
   public void execute() {
-    m_intake.TalonFlex(-0.5);;
+    m_pivot.pivot(ControlMap.gunner.getRawAxis(4));
+   
   }
 
   // Called once the command ends or is interrupted.
   @Override
-  public void end(boolean interrupted) {
-    m_intake.TalonFlex(0);
-  }
+  public void end(boolean interrupted) {}
+
 
   // Returns true when the command should end.
   @Override
   public boolean isFinished() {
-    if(m_intake.upperLimitGoesOff()){
-      return true;
-    }
+
     return false;
   }
 }
