+++ src/main/java/frc/robot/commands/Climb/RetractClimb.java
@@ -2,26 +2,24 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
+package frc.robot.commands.Climb;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
-import frc.robot.subsystems.Intake;
+import frc.robot.subsystems.Climb;
 
-public class TopClose extends CommandBase {
-  /** Creates a new TopClose. */
-  private final Intake m_intake;
-  public TopClose(Intake intake) {
-    // Use addRequirements() here to declare subsystem dependencies.
-    m_intake = intake;
-    addRequirements(intake);
-  }
+public class RetractClimb extends CommandBase {
+  private final Climb m_climb;
 
-  // Called when the command is initially scheduled.
-  @Override
-  public void initialize() {
-    //m_intake.TopServoClose();
-  }
-
+    //Constructor
+    public RetractClimb(Climb climb){
+        m_climb = climb;
+        addRequirements(climb);
+    }
+    //Extend the pivot arm
+    @Override
+    public void initialize(){
+        m_climb.retractArm();
+    }
   // Called every time the scheduler runs while the command is scheduled.
   @Override
   public void execute() {}
