+++ src/main/java/frc/robot/commands/Climb/ToggleHooks.java
@@ -1,20 +1,25 @@
-package frc.robot.commands;
+// Copyright (c) FIRST and other WPILib contributors.
+// Open Source Software; you can modify and/or share it under the terms of
+// the WPILib BSD license file in the root directory of this project.
+
+package frc.robot.commands.Climb;
+
 import edu.wpi.first.wpilibj2.command.CommandBase;
 import frc.robot.subsystems.Climb;
 
-public class RetractArm extends CommandBase {
-  /** Creates a new RetractArm. */
+public class ToggleHooks extends CommandBase {
+  /** Creates a new ToggleHooks. */
   private final Climb m_climb;
-  public RetractArm(Climb climb) {
-    // Use addRequirements() here to declare subsystem dependencies.
+  public ToggleHooks(Climb climb) {
     m_climb = climb;
+    // Use addRequirements() here to declare subsystem dependencies.
     addRequirements(climb);
   }
 
   // Called when the command is initially scheduled.
   @Override
   public void initialize() {
-    m_climb.retractArm();
+    m_climb.ServoOpen();
   }
 
   // Called every time the scheduler runs while the command is scheduled.
@@ -23,7 +28,9 @@ public class RetractArm extends CommandBase {
 
   // Called once the command ends or is interrupted.
   @Override
-  public void end(boolean interrupted) {}
+  public void end(boolean interrupted) {
+    m_climb.ServoClose();
+  }
 
   // Returns true when the command should end.
   @Override
