+++ src/main/java/frc/robot/commands/Intake/ArmReverse.java
@@ -2,15 +2,16 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
+package frc.robot.commands.Intake;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
 import frc.robot.subsystems.Intake;
 
-public class BottomOpen extends CommandBase {
+public class ArmReverse extends CommandBase {
   /** Creates a new TopClose. */
+  
   private final Intake m_intake;
-  public BottomOpen(Intake intake) {
+  public ArmReverse(Intake intake) {
     // Use addRequirements() here to declare subsystem dependencies.
     m_intake = intake;
     addRequirements(intake);
@@ -19,20 +20,27 @@ public class BottomOpen extends CommandBase {
   // Called when the command is initially scheduled.
   @Override
   public void initialize() {
-   // m_intake.BottomServoClose();
+    
   }
 
   // Called every time the scheduler runs while the command is scheduled.
   @Override
-  public void execute() {}
+  public void execute() {
+    m_intake.TalonFlex(-0.5);;
+  }
 
   // Called once the command ends or is interrupted.
   @Override
-  public void end(boolean interrupted) {}
+  public void end(boolean interrupted) {
+    m_intake.TalonFlex(0);
+  }
 
   // Returns true when the command should end.
   @Override
   public boolean isFinished() {
+    if(m_intake.upperLimitGoesOff()){
+      return true;
+    }
     return false;
   }
 }
