+++ src/main/java/frc/robot/commands/Intake/BottomClose.java
@@ -2,25 +2,24 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
+package frc.robot.commands.Intake;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
-import frc.robot.subsystems.Climb;
+import frc.robot.subsystems.Intake;
 
-public class HookOn extends CommandBase {
-  /** Creates a new HookOn. */
-  private final Climb m_climb;
-
-  public HookOn(Climb climb) {
+public class BottomClose extends CommandBase {
+  /** Creates a new TopClose. */
+  private final Intake m_intake;
+  public BottomClose(Intake intake) {
     // Use addRequirements() here to declare subsystem dependencies.
-    m_climb = climb;
-    addRequirements(climb);
+    m_intake = intake;
+    addRequirements(intake);
   }
 
   // Called when the command is initially scheduled.
   @Override
   public void initialize() {
-    m_climb.ServoClose();
+    //m_intake.BottomServoClose();
   }
 
   // Called every time the scheduler runs while the command is scheduled.
