+++ src/main/java/frc/robot/commands/Intake/BottomOpen.java
@@ -2,42 +2,37 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
-import edu.wpi.first.wpilibj2.command.CommandBase;
-import frc.robot.subsystems.Climb;
+package frc.robot.commands.Intake;
 
-public class ServoOpen extends CommandBase {
-  /** Creates a new Pivot. */
-  private final Climb m_servoOpen;
+import edu.wpi.first.wpilibj2.command.CommandBase;
+import frc.robot.subsystems.Intake;
+
+public class BottomOpen extends CommandBase {
+  /** Creates a new TopClose. */
+  private final Intake m_intake;
+  public BottomOpen(Intake intake) {
     // Use addRequirements() here to declare subsystem dependencies.
-    //Constructor
-    public ServoOpen(Climb ServoOpen){
-        m_servoOpen = ServoOpen;
-        addRequirements(ServoOpen);
+    m_intake = intake;
+    addRequirements(intake);
   }
 
   // Called when the command is initially scheduled.
   @Override
   public void initialize() {
-      m_servoOpen.ServoOpen();
-      
-
+   // m_intake.BottomServoClose();
   }
 
   // Called every time the scheduler runs while the command is scheduled.
   @Override
-  public void execute() {
-
-  }
+  public void execute() {}
 
   // Called once the command ends or is interrupted.
   @Override
-  public void end(boolean interrupted) {
-  }
+  public void end(boolean interrupted) {}
 
   // Returns true when the command should end.
-  @Override 
+  @Override
   public boolean isFinished() {
-      return true;
+    return false;
   }
 }
