--- src/main/java/frc/robot/commands/ToggleClimb.java
@@ -1,40 +1,0 @@
-// Copyright (c) FIRST and other WPILib contributors.
-// Open Source Software; you can modify and/or share it under the terms of
-// the WPILib BSD license file in the root directory of this project.
-
-package frc.robot.commands;
-
-import edu.wpi.first.wpilibj2.command.CommandBase;
-import frc.robot.subsystems.Climb;
-
-public class ToggleClimb extends CommandBase {
-  /** Creates a new ToggleClimb. */
-  private final Climb m_climb;
-  public ToggleClimb(Climb climb) {
-    m_climb = climb;
-    // Use addRequirements() here to declare subsystem dependencies.
-    addRequirements(climb);
-  }
-
-  // Called when the command is initially scheduled.
-  @Override
-  public void initialize() {
-    m_climb.extendArm();
-  }
-
-  // Called every time the scheduler runs while the command is scheduled.
-  @Override
-  public void execute() {}
-
-  // Called once the command ends or is interrupted.
-  @Override
-  public void end(boolean interrupted) {
-    m_climb.retractArm();
-  }
-
-  // Returns true when the command should end.
-  @Override
-  public boolean isFinished() {
-    return false;
-  }
-}
