--- src/main/java/frc/robot/commands/TopOpen.java
@@ -1,38 +1,0 @@
-// Copyright (c) FIRST and other WPILib contributors.
-// Open Source Software; you can modify and/or share it under the terms of
-// the WPILib BSD license file in the root directory of this project.
-
-package frc.robot.commands;
-
-import edu.wpi.first.wpilibj2.command.CommandBase;
-import frc.robot.subsystems.Intake;
-
-public class TopOpen extends CommandBase {
-  /** Creates a new TopClose. */
-  private final Intake m_intake;
-  public TopOpen(Intake intake) {
-    // Use addRequirements() here to declare subsystem dependencies.
-    m_intake = intake;
-    addRequirements(intake);
-  }
-
-  // Called when the command is initially scheduled.
-  @Override
-  public void initialize() {
-    //m_intake.TopServoOpen();
-  }
-
-  // Called every time the scheduler runs while the command is scheduled.
-  @Override
-  public void execute() {}
-
-  // Called once the command ends or is interrupted.
-  @Override
-  public void end(boolean interrupted) {}
-
-  // Returns true when the command should end.
-  @Override
-  public boolean isFinished() {
-    return false;
-  }
-}
